+++ modules/graphics/tests/test_figure.m
@@ -53,11 +53,13 @@ REF = {'AlphaMap';
 'Name';
 'NextPlot';
 'Number';
+'NumberTitle';
 'Parent';
 'Position';
 'Tag';
 'Type';
-'UserData'};
+'UserData';
+'Visible'};
 assert_isequal(p, REF);
 %=============================================================================
 close all