+++ CHANGELOG.md
@@ -10,9 +10,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - `drawnow`: Update figures and process callbacks
 - `DrawLater` property added to `figure` graphics object.
+- `interp1` linear interpolation 1D.
+- [736](http://github.com/Nelson-numerical-software/nelson/issues/736): `bone`, `cool`, `copper`, `hot`, `jet`, `pink`, `turbo`, `viridis`, `white` colormaps.
 - `CHANGELOG` 0.7.x family.
+- `Visible` property to `figure` graphics object.
+- [809](http://github.com/Nelson-numerical-software/nelson/issues/809): `NumberTitle` property to `figure` graphics object.
+- `AlphaMap` and `Colormap` properties added to `Axes` graphics object.
 
 ### Fixed
+
+- [#823](http://github.com/Nelson-numerical-software/nelson/issues/823): default LineStyle for a line was wrong with marker.
 
 - warnings detected by CodeQL.